Modificacion de la base de datos (casi definitiva), modificacion de algunos tipos de datos

+++ TPC_Equipo-6A/Dominio/Categoria.cs
@@ -8,7 +8,7 @@ namespace dominio
 {
     public class Categoria
     {
-        public long IdCategoria { get; set; }
+        public int IdCategoria { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public bool Estado { get; set; }
